# PERF: Improve perf of matrix.sum_rows