same impl as #115.

### Before

```
test linalg::matrix::mat_mean_cols                 ... bench:       8,800 ns/iter (+/- 200)
test linalg::matrix::mat_mean_rows                 ... bench:      18,792 ns/iter (+/- 2,265)
test linalg::matrix::mat_sum_cols                  ... bench:       4,915 ns/iter (+/- 1,770)
test linalg::matrix::mat_sum_rows                  ... bench:      12,735 ns/iter (+/- 1,606)
test linalg::matrix::mat_sum_rows_and_cols_128_100 ... bench:      27,507 ns/iter (+/- 9,377)
```

### After 

```
test linalg::matrix::mat_mean_cols                 ... bench:       9,528 ns/iter (+/- 3,713)
test linalg::matrix::mat_mean_rows                 ... bench:      15,885 ns/iter (+/- 4,485)
test linalg::matrix::mat_sum_cols                  ... bench:       4,939 ns/iter (+/- 1,895)
test linalg::matrix::mat_sum_rows                  ... bench:       8,343 ns/iter (+/- 1,591)
test linalg::matrix::mat_sum_rows_and_cols_128_100 ... bench:      21,615 ns/iter (+/- 3,500)
```