@tafia 2016-12-25
I do not understand why it is faster. Is this some new compiler optimization ?