@AtheMathmo 2016-12-25
I also am a little clueless. I know that the compiler can often do some magic with slices - it aids cache access and vectorization. I'm not sure why this particular example gives a speed up though.

I am going to try it on my own architecture to see if I reproduce the results.