@AtheMathmo 2016-12-27
Just pinging to let you know there are merge conflicts here. Once resolved I'll check out and see if I can reproduce the performance gain. If yes, will merge this happily.

Do you have any idea why this gives a performance improvement @sinhrks ?