@sinhrks 2016-12-27
I **assume** ``row_sums`` is now updated in-place, rather than allocate every-time via ``vec_bin_op``.