@AtheMathmo 2016-12-27
Ah that does make sense.

Could you humor me and try using `utils::in_place_vec_bin_op`? So that you update the row sum in place and use explicit vectorization (which in the past has worked better than zip iteration).