@sinhrks 2016-12-27
updated to use ``in_place_vec_bin_op``, which can achieve almost the same perf. 

```
test linalg::matrix::mat_mean_cols                 ... bench:       8,814 ns/iter (+/- 3,091)
test linalg::matrix::mat_mean_rows                 ... bench:      14,800 ns/iter (+/- 4,207)
test linalg::matrix::mat_sum_cols                  ... bench:       4,948 ns/iter (+/- 2,381)
test linalg::matrix::mat_sum_rows                  ... bench:       8,419 ns/iter (+/- 997)
test linalg::matrix::mat_sum_rows_and_cols_128_100 ... bench:      22,440 ns/iter (+/- 9,303)
```