@AtheMathmo 2016-12-27
I didn't mean that you had to make the change, sorry if that was unclear. It does look like it is more-or-less the same though, which is good. Thanks for checking.

I'm happy to merge this now. How about you @tafia ?