@sinhrks 2016-12-27
@AtheMathmo Understood. I intend to use common in-place logic as we might find better in-place impl. I'm going to change #115 to use ``in_place_vec_bin_op`` if OK.