@tafia 2016-12-27
Yes of course! Any performance gain is very welcomed. Both the explicit loop and the helper function look good to me.